+++ pkg/network/anonymity/examples/echo/construct.go
@@ -91,6 +91,7 @@ func newNode(serviceName, address string) anonymity.INode {
 				}),
 				asymmetric.NewRSAPrivKey(keySize),
 			),
+			asymmetric.NewRSAPrivKey(keySize).GetPubKey(),
 		),
 		asymmetric.NewListPubKeys(),
 	)