+++ pkg/network/anonymity/queue/examples/queue/main.go
@@ -33,6 +33,7 @@ func main() {
 			}),
 			asymmetric.NewRSAPrivKey(1024),
 		),
+		asymmetric.NewRSAPrivKey(1024).GetPubKey(),
 	)
 
 	ctx, cancel := context.WithCancel(context.Background())